feat(COR-1954): Consolidated metadataprops type into seperate type file
